+++ scripts/clean-modules/index.js
@@ -2,3 +2,4 @@
 const shell = require('shelljs');
 
 shell.rm('-rf', 'node_modules');
+shell.rm('-rf', 'lib');