+++ packages/terra-application-utility/examples/test-examples/DefaultApplicationUtility.jsx
@@ -1,0 +1,5 @@
+import React from 'react';
+
+import ApplicationUtility from '../../lib/ApplicationUtility';
+
+export default () => <ApplicationUtility />;