+++ packages/terra-layout/CHANGELOG.md
@@ -8,6 +8,10 @@ Unreleased
 ------------------
 ### Added
 * Debounce to resize of layout
+* Use display none on the hover section when hidden.
+
+###Changed
+* Updating tests to point to new terra-framework-site location.
 
 1.0.0 - (November 17, 2017)
 ------------------