+++ packages/terra-layout/CHANGELOG.md
@@ -4,6 +4,15 @@ Changelog
 Unreleased
 ----------
 
+1.1.0 - (January 23, 2018)
+------------------
+### Added
+* Debounce to resize of layout
+* Use display none on the hover section when hidden.
+
+###Changed
+* Updating tests to point to new terra-framework-site location.
+
 1.0.0 - (November 17, 2017)
 ------------------
 Initial stable release